+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Autoscaler Change Log
 ------------------------
 
 - Fix predict policy: do not mess original workflow tracker
+- Fix missing provisioning time in predict policy
 
 0.5.0 October 27, 2020
 ------------------------